+++ core/execution/engine.go
@@ -42,6 +42,9 @@ var (
 
 	// ErrSuccessorMarketDoesNotExists is returned when SucceedMarket call is made with an invalid successor market ID.
 	ErrSuccessorMarketDoesNotExist = errors.New("successor market does not exist")
+
+	// ErrParentMarketNotEnactedYEt is returned when trying to enact a successor market that is still in proposed state.
+	ErrParentMarketNotEnactedYet = errors.New("parent market in proposed state, can't enact successor")
 )
 
 // Engine is the execution engine.
@@ -78,6 +81,8 @@ type Engine struct {
 	successors      map[string][]string
 	isSuccessor     map[string]string
 	successorWindow time.Duration
+	// map tracking enacted successor markets for given parent markets for a given parent
+	pendingSuccessors map[string]string
 }
 
 type netParamsValues struct {
@@ -150,6 +155,7 @@ func NewEngine(
 		marketCPStates:        map[string]*types.CPMarketState{},
 		successors:            map[string][]string{},
 		isSuccessor:           map[string]string{},
+		pendingSuccessors:     map[string]string{},
 	}
 
 	// set the eligibility for proposer bonus checker
@@ -260,8 +266,8 @@ func (e *Engine) StartOpeningAuction(ctx context.Context, marketID string) error
 	return nil
 }
 
-func (e *Engine) SucceedMarket(ctx context.Context, successor, parent string, insuranceFraction num.Decimal) error {
-	return e.succeedOrRestore(ctx, successor, parent, insuranceFraction, false)
+func (e *Engine) SucceedMarket(ctx context.Context, successor, parent string) error {
+	return e.succeedOrRestore(ctx, successor, parent, false)
 }
 
 func (e *Engine) restoreOwnState(ctx context.Context, mID string) (bool, error) {
@@ -292,7 +298,7 @@ func (e *Engine) restoreOwnState(ctx context.Context, mID string) (bool, error) 
 	return false, nil
 }
 
-func (e *Engine) succeedOrRestore(ctx context.Context, successor, parent string, insuranceFraction num.Decimal, restore bool) error {
+func (e *Engine) succeedOrRestore(ctx context.Context, successor, parent string, restore bool) error {
 	mkt, ok := e.markets[successor]
 	if !ok {
 		// this can happen if a proposal vote closed, but the proposal had an enactment time in the future.
@@ -302,7 +308,7 @@ func (e *Engine) succeedOrRestore(ctx context.Context, successor, parent string,
 		return ErrMarketDoesNotExist
 	}
 	// if this is a market restore, first check to see if there is some state already
-	pm, ok := e.GetMarket(parent, true)
+	_, ok = e.GetMarket(parent, true)
 	if !ok && !restore {
 		// a successor market that has passed the vote, but the parent market either already was succeeded
 		// or the proposal vote closed when the parent market was still around, but it wasn't enacted until now
@@ -311,11 +317,7 @@ func (e *Engine) succeedOrRestore(ctx context.Context, successor, parent string,
 		mkt.ResetParentIDAndInsurancePoolFraction()
 		return nil
 	}
-	var parentM *types.Market
-	if ok {
-		parentM = &pm
-	}
-	parentState, sok := e.marketCPStates[parent]
+	_, sok := e.marketCPStates[parent]
 	// restoring a market, but no state of the market nor parent market exists. Treat market as parent.
 	if restore && !sok && !ok {
 		// restoring a market, but the market state and parent market both are missing
@@ -324,25 +326,15 @@ func (e *Engine) succeedOrRestore(ctx context.Context, successor, parent string,
 	}
 	// if parent market is active, mark as succeeded
 	if pmo, ok := e.markets[parent]; ok {
+		// succeeding a parent market before it was enacted is not allowed
+		if pmo.Mkt().State == types.MarketStateProposed {
+			e.RejectMarket(ctx, successor)
+			return ErrParentMarketNotEnactedYet
+		}
 		// mark as succeeded so the state is excluded from checkpoint data
 		pmo.SetSucceeded()
-		parentM = pmo.Mkt()
-		// it may be possible that there is no CP state for this market yet
-		if !sok {
-			// no cp state (yet), get the state from the parent market directly
-			parentState = pmo.GetCPState()
-		}
 	}
-	if !insuranceFraction.IsZero() {
-		lm := e.collateral.SuccessorInsuranceFraction(ctx, successor, parent, mkt.GetSettlementAsset(), insuranceFraction)
-		e.broker.Send(events.NewLedgerMovements(ctx, []*types.LedgerMovement{lm}))
-	}
-	// pass in the ELS and the like
-	mkt.InheritParent(ctx, parentState)
-	// if the parent market is in trading terminated state, then just remove it from the CP states
-	if !restore && parentM != nil && parentM.State == types.MarketStateSettled {
-		delete(e.marketCPStates, parent)
-	}
+	e.pendingSuccessors[parent] = successor
 	// successor market set up accordingly, clean up the state
 	// first reject all pending successors proposed for the same parent
 	for _, pending := range e.successors[parent] {
@@ -390,7 +382,7 @@ func (e *Engine) RestoreMarket(ctx context.Context, marketConfig *types.Market) 
 	}
 	// this is a successor market, handle accordingly
 	if pid := marketConfig.ParentMarketID; len(pid) > 0 {
-		return e.succeedOrRestore(ctx, marketConfig.ID, pid, marketConfig.InsurancePoolFraction, true)
+		return e.succeedOrRestore(ctx, marketConfig.ID, pid, true)
 	}
 	return nil
 }
@@ -884,7 +876,27 @@ func (e *Engine) OnTick(ctx context.Context, t time.Time) {
 				logging.MarketID(id))
 			toDelete = append(toDelete, id)
 		}
-		if !mkt.IsSucceeded() {
+		mdef := mkt.Mkt()
+		if _, ok := e.pendingSuccessors[mdef.ParentMarketID]; ok && mdef.State != types.MarketStatePending && mdef.State != types.MarketStateProposed {
+			// the successor market has left opening auction, get the state if possible
+			if cps, ok := e.marketCPStates[mdef.ParentMarketID]; ok {
+				if !mdef.InsurancePoolFraction.IsZero() {
+					lm := e.collateral.SuccessorInsuranceFraction(ctx, id, mdef.ParentMarketID, mkt.GetSettlementAsset(), mdef.InsurancePoolFraction)
+					if lm != nil {
+						e.broker.Send(events.NewLedgerMovements(ctx, []*types.LedgerMovement{lm}))
+					}
+				}
+				mkt.InheritParent(ctx, cps)
+				// successor market done
+				delete(e.marketCPStates, mdef.ParentMarketID)
+			} else {
+				// parent market state is long gone, this market is not a successor
+				mkt.ResetParentIDAndInsurancePoolFraction()
+			}
+			// market left opening auction, if no state was inherited/transferred now, it never will be
+			delete(e.pendingSuccessors, mdef.ParentMarketID)
+		}
+		if _, ok := e.pendingSuccessors[id]; ok || !mkt.IsSucceeded() {
 			// update the market state used to set the successor market accordingly
 			// if the market was closed, the checkpoint data will include the full market definition
 			cps := mkt.GetCPState()
@@ -892,7 +904,8 @@ func (e *Engine) OnTick(ctx context.Context, t time.Time) {
 			cps.TTL = t.Add(e.successorWindow)
 			e.marketCPStates[id] = cps
 		} else {
-			// just in case it's still around -> remove the state
+			// just in case it's still around -> remove the state, this would mean there is no pending successor, and the market is
+			// marked as succeeded, this state should not be here
 			delete(e.marketCPStates, id)
 		}
 		evts = append(evts, events.NewMarketDataEvent(ctx, mkt.GetMarketData()))
@@ -904,6 +917,7 @@ func (e *Engine) OnTick(ctx context.Context, t time.Time) {
 	}
 	// clear slice
 	toDelete = toDelete[:]
+	// find state that should expire
 	for id, cpm := range e.marketCPStates {
 		if !cpm.TTL.After(t) {
 			toDelete = append(toDelete, id)
@@ -911,6 +925,13 @@ func (e *Engine) OnTick(ctx context.Context, t time.Time) {
 	}
 	for _, id := range toDelete {
 		delete(e.marketCPStates, id)
+		if s, ok := e.pendingSuccessors[id]; ok {
+			// parent market expired, remove parent ID
+			if mkt, ok := e.markets[s]; ok {
+				mkt.ResetParentIDAndInsurancePoolFraction()
+			}
+		}
+		delete(e.pendingSuccessors, id)
 	}
 
 	timer.EngineTimeCounterAdd()
@@ -928,6 +949,15 @@ func (e *Engine) GetMarketState(mktID string) (types.MarketState, error) {
 		return types.MarketStateUnspecified, types.ErrInvalidMarketID
 	}
 	return mkt.GetMarketState(), nil
+}
+
+func (e *Engine) IsSucceeded(mktID string) bool {
+	if mkt, ok := e.markets[mktID]; ok {
+		return mkt.IsSucceeded()
+	}
+	// checking marketCPStates is pointless. The parent market could not be found to validate the proposal, so it will be rejected outright
+	// if the market is no longer in e.markets, it will be set in marketCPStates, and therefore the successor proposal must be accepted.
+	return false
 }
 
 func (e *Engine) GetMarketData(mktID string) (types.MarketData, error) {