+++ core/execution/engine_test.go
@@ -2,6 +2,7 @@ package execution_test
 
 import (
 	"context"
+	"encoding/hex"
 	"fmt"
 	"testing"
 	"time"
@@ -10,6 +11,7 @@ import (
 	"code.vegaprotocol.io/vega/core/events"
 	"code.vegaprotocol.io/vega/core/oracles"
 	"code.vegaprotocol.io/vega/core/types"
+	vgcontext "code.vegaprotocol.io/vega/libs/context"
 	"code.vegaprotocol.io/vega/libs/num"
 
 	"github.com/golang/mock/gomock"
@@ -23,7 +25,7 @@ func TestMarketSuccession(t *testing.T) {
 	mkt := getMarketConfig()
 	mkt.ID = "parentID"
 	// sendCount, batchCount := 0, 0
-	ctx := context.Background()
+	ctx := vgcontext.WithTraceID(context.Background(), hex.EncodeToString([]byte("0deadbeef")))
 	// for now, we don't care about this much
 	// exec.epoch.EXPECT().NotifyOnEpoch(gomock.Any(), gomock.Any()).AnyTimes()
 	exec.asset.EXPECT().Get(gomock.Any()).AnyTimes().DoAndReturn(func(asset string) (*assets.Asset, error) {
@@ -38,20 +40,14 @@ func TestMarketSuccession(t *testing.T) {
 		return a, nil
 	})
 	// this is to propose the parent market and the 2 successors, and starting the opening auction for the eventual successor
-	exec.broker.EXPECT().Send(gomock.Any()).Times(10)
-	// exec.broker.EXPECT().Send(gomock.Any()).AnyTimes().Do(func(_ events.Event) {
-	// sendCount++
-	// })
-	// exec.broker.EXPECT().SendBatch(gomock.Any()).AnyTimes().Do(func(_ []events.Event) {
-	// batchCount++
-	// })
+	seen := false
+	exec.broker.EXPECT().Send(gomock.Any()).Times(19).Do(func(e events.Event) {
+		if e.Type() == events.MarketUpdatedEvent {
+			seen = true
+		}
+	})
+	exec.broker.EXPECT().SendBatch(gomock.Any()).AnyTimes()
 	exec.collateral.EXPECT().AssetExists(gomock.Any()).AnyTimes().Return(true)
-	// exec.collateral.EXPECT().AssetExists(gomock.Any()).AnyTimes().DoAndReturn(func(a string) bool {
-	// if s, ok := knownAssets[a]; ok && s != nil {
-	// return true
-	// }
-	// return false
-	// })
 	exec.collateral.EXPECT().CreateMarketAccounts(gomock.Any(), gomock.Any(), gomock.Any()).AnyTimes()
 	exec.oracle.EXPECT().Subscribe(gomock.Any(), gomock.Any(), gomock.Any()).AnyTimes().Return(oracles.SubscriptionID(0), func(_ context.Context, _ oracles.SubscriptionID) {})
 	exec.statevar.EXPECT().RegisterStateVariable(gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any(), gomock.Any()).AnyTimes()
@@ -83,6 +79,10 @@ func TestMarketSuccession(t *testing.T) {
 	// when enacting a successor market, a lot of stuff happens:
 
 	// Transfer insurance pool fraction
+	acc := &types.Account{
+		Balance: num.UintZero(),
+	}
+	exec.collateral.EXPECT().GetMarketLiquidityFeeAccount(gomock.Any(), gomock.Any()).AnyTimes().Return(acc, nil)
 	exec.collateral.EXPECT().SuccessorInsuranceFraction(ctx, child1.ID, child1.ParentMarketID, gomock.Any(), child1.InsurancePoolFraction).Times(1).Return(&types.LedgerMovement{})
 	// which in turn emits an event with ledger movements
 	exec.broker.EXPECT().Send(gomock.Any()).Times(1).Do(func(evt events.Event) {
@@ -91,19 +91,20 @@ func TestMarketSuccession(t *testing.T) {
 	})
 	// we get the parent market state to pass in the ELS and stuff:
 	exec.collateral.EXPECT().GetInsurancePoolBalance(child1.ParentMarketID, gomock.Any()).Times(1).Return(num.NewUint(100), true) // the balance doesn't matter for this test
+	exec.collateral.EXPECT().GetInsurancePoolBalance(child1.ID, gomock.Any()).Times(1).Return(num.NewUint(50), true)              // the balance doesn't matter for this test
 	// Any accounts associated with the now rejected successor market will be removed
 	exec.collateral.EXPECT().ClearMarket(ctx, child2.ID, gomock.Any(), gomock.Any()).Times(1).Return(nil, nil)
 	// statevars associated with the rejected successor market are unregistered
 	exec.statevar.EXPECT().UnregisterStateVariable(gomock.Any(), child2.ID).AnyTimes()
 	// the other succesor markets are rejected and removed, which emits market update events
-	exec.broker.EXPECT().Send(gomock.Any()).Times(1).Do(func(evt events.Event) {
-		require.Equal(t, events.MarketUpdatedEvent, evt.Type())
-	})
 	// set parent market to be settled
 	err = exec.StartOpeningAuction(ctx, child1.ID)
 	require.NoError(t, err)
 	mkt.State = types.MarketStateSettled
+	child1.State = types.MarketStateActive
 	// start opening auction for the successor market
-	err = exec.SucceedMarket(ctx, child1.ID, child1.ParentMarketID, child1.InsurancePoolFraction)
+	err = exec.SucceedMarket(ctx, child1.ID, child1.ParentMarketID)
 	require.NoError(t, err)
+	exec.OnTick(ctx, time.Now())
+	require.True(t, seen)
 }